Move update switch copy-paste logic to switch.c
