+++ src/ui/switch.h
@@ -47,5 +47,6 @@ bool switch_mright(UISWITCH *s);
 bool switch_mwheel(UISWITCH *s, int height, double d, bool smooth);
 bool switch_mup(UISWITCH *s);
 bool switch_mleave(UISWITCH *s);
+void switch_update(UISWITCH *s);
 
 #endif
